Gamer lights might work now.
